feat(i18n): add internationalization support with Arabic and English - needs to fix in future
- Implement i18n infrastructure with locale detection and translation loading
- Add Arabic and English translation files for all UI text
- Create TranslationProvider component for client-side translations
- Update Navigation and Footer components to use translations
- Add middleware for locale detection and routing
- Move all pages to locale-based routing structure
- Clean up animation styles in HeroSection
