+++ src/messages/ar.json
@@ -1,0 +1,26 @@
+{
+  "common": {
+    "name": "سالم الجبالي",
+    "title": "مدير تقني وخبير DevOps",
+    "downloadCV": "السيرة الذاتية",
+    "allRightsReserved": "جميع الحقوق محفوظة."
+  },
+  "nav": {
+    "home": "الرئيسية",
+    "about": "حول",
+    "projects": "المشاريع",
+    "certifications": "الشهادات",
+    "contact": "تواصل"
+  },
+  "home": {
+    "heroSubtitle": "تحويل الأعمال من خلال الأتمتة والبنية التحتية السحابية والقيادة التقنية",
+    "ctaTitle": "هل أنت مهتم بتحويل البنية التحتية التقنية الخاصة بك؟",
+    "ctaDescription": "دعنا نناقش كيف يمكنني مساعدة مؤسستك في تحقيق أهدافها.",
+    "ctaButton": "لنتحدث"
+  },
+  "footer": {
+    "quickLinks": "روابط سريعة",
+    "contactInfo": "معلومات التواصل",
+    "location": "ليبيا"
+  }
+}
